Copy over all the modules that should be part of Kimono from yui3-gallery and yui3
While copying them over, I made some modifications that I'd like some feedback on:

- Changed module names to match what's on the website. These names are cleaner
and don't have the "css" prefix cruft.
- Removed the meta/ directory from the modules. I didn't find a need for this because
presumably, people will use this without being part of YUI's (and Loader's) ecosystem.
If anything, I felt that there should be a Grunt script that copies over this CSS into
YUI and has a meta/ directory in a YUI port of this Kimono stuff.
- Renamed cssnormalize to base.
- Did not include a build/ directory, but it seems like shifter needs one at the top-level
to put the built files in. Locally, I have build/ and src/, but I just included src/, and
added build/ to the .gitignore