Add `[hidden]` selector counteract all `display` declarations
Anytime a `display` declaration is used in Pure with a value other than
`none`, we need to add an additional rule with similar selectors
(appended with `[hidden]`) with a `display: none;` declaration. This
allows the `hidden` HTML attribute to work properly on elements styled
by Pure.

The following button will be hidden from view with these changes:

    <button class="pure-button" hidden>No showy</button>

Fixes #177