Add `[hidden]` selector to Base to counteract all `display` declarations
Anytime a Pure style rule uses the `display` declaration with a value
other than `none`, using the `hidden` HTML attribute on that element
won't have the desired affect of hiding that element. To counteract
this, A new rule has been added to Base to force
`display: none !important;` on elements that use the `hidden` HTML
attribute.

The following button will be hidden from view with these changes:

    <button class="pure-button" hidden>No showy</button>

Fixes #177