Fixed grouped input square corners special case
When there is only one input in a .pure-group, it had square top corners. Added a :first-child:last-child rule to have round corners as it should.

This is useful when you want grouped input fields, but where the input elements might be dynamically added to the group, so you start with one input, but more are added later.